:sparkles: Implement straight logic
Implement a check to determine whether the hand is a stright.

Consider the options when a T, J, Q, K or A are in the hand. The
values are placed in a vector and if the sorted result matches
these options, return true.

Else, loop through the values and check that they are consecutive.

For the straight flush, refactor that to return true iff straight &
flush are true