:triangular_ruler: Hit it with the math hammer :hammer:
Quite an interesting implementation here. Decided to keep both
the numberator and denominator seperate, and use some basic
fractions knowledge to do the calculations.

Unfortunatley, the fractions became veey large very quckly so we
needed to chop off some digits. This means that we lost presicion
quite quickly and needed to be careful. In any case the decision
of > 1 000 000 000 /= 10 000 does the trick for this question.

Runs in a couple miliseconds. Which is good.