:sparkles: Another prime number question
This one we get a few quick wins by early returning at each level if one
of those values is not prime. The difficulty with these brute force
methods is that one needs to know where to set the boundries.

Once the first solution didn't work, the logical next step is
to change the first value by a bit. Else just let it run
for a very long time and hope for the best. Knowing
rough guidelines means this runs in 700ms, not setting them
changes it to 4 seconds.